splitting code for readability
