+++ src/components/navigation/constants.ts
@@ -1,0 +1,37 @@
+export const TRANSITION_DURATION = 300;
+export const PADDING_LEVELS = {
+  default: {
+    left: "0.675rem",
+    top: "0.125rem",
+    bottom: "0.125rem",
+  },
+  level0: {
+    left: "0.75rem",
+    top: "0.25rem",
+    bottom: "0.125rem",
+  },
+};
+
+export const TEXT_COLORS = {
+  default: "text-gray-800",
+  selected: "text-blue-500",
+  childSelected: "text-gray-800",
+  orange: "text-orange-500",
+  orangeLight: "text-orange-100",
+  orangeHover: "text-orange-300",
+  blueLight: "text-blue-200",
+  blueHover: "text-blue-400",
+};
+
+export const FONT_SIZES = {
+  xl: "text-xl",
+  base: "text-base",
+  small: "text-[15px]",
+};
+
+export const FONT_WEIGHTS = {
+  light: "font-tuner-light",
+  normal: "font-sans",
+  medium: "font-[500]",
+  bold: "font-bold",
+};
