+++ src/components/navigation/type.ts
@@ -1,0 +1,11 @@
+export interface DocsNavProps {
+  navItems: any;
+}
+
+export interface NavTitleProps {
+  level: number;
+  selected: boolean;
+  childSelected?: boolean;
+  children: React.ReactNode | React.ReactNode[];
+  onClick?: () => void;
+}
